Make druid a separate indexed column on events
This creates a migration and adds a handler to set up a top-level
field on Event objects for the druid, so that it can be set and
queried without resorting to pulling it out of the properties
text field.

The changes is intended to address performance concerns for
objects with lots of non-unique events, where we need to do
queries by druid and then unique them by event id.